requirements.txt
pycryptodome
google-cloud-bigquery
requests
